+++ scripts/run-test-migrations.php
@@ -136,6 +136,7 @@ $requiredTables = [
     'content_types',
     'entries',
     'entry_redirects',
+    'entry_schedules',
 ];
 
 $dsn = sprintf(